worker/requirements.txt
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
Pillow==10.2.0
